requirements.txt
# Core LangChain (using 0.1.20 - stable version with working imports)
langchain==0.1.20
langchain-core==0.1.52
langchain-community==0.0.38

# FREE LLM Provider (Google Gemini)
langchain-google-genai==1.0.1

# Vector Store and Embeddings
faiss-cpu>=1.8.0
sentence-transformers>=2.5.1

# PDF Processing
pypdf>=4.1.0

# Web Interface
streamlit>=1.32.0

# Utilities
python-dotenv>=1.0.0
